Move ClassScopeHandler to a separate class
