+++ packages/babel-parser/src/parser/index.js
@@ -7,6 +7,7 @@ import { getOptions } from "../options";
 import StatementParser from "./statement";
 import { SCOPE_ASYNC, SCOPE_PROGRAM } from "../util/scopeflags";
 import ScopeHandler from "../util/scope";
+import ClassScopeHandler from "../util/class-scope";
 
 export type PluginsMap = Map<string, { [string]: any }>;
 
@@ -25,6 +26,7 @@ export default class Parser extends StatementParser {
     this.options = options;
     this.inModule = this.options.sourceType === "module";
     this.scope = new ScopeHandler(this.raise.bind(this), this.inModule);
+    this.classScope = new ClassScopeHandler(this.raise.bind(this));
     this.plugins = pluginsMap(this.options.plugins);
     this.filename = options.sourceFilename;
   }
