+++ packages/babel-parser/src/tokenizer/state.js
@@ -77,9 +77,6 @@ export default class State {
   soloAwait: boolean = false;
   inFSharpPipelineDirectBody: boolean = false;
 
-  // Check whether we are in a (nested) class or not.
-  classLevel: number = 0;
-
   // Labels in scope.
   labels: Array<{
     kind: ?("loop" | "switch"),
